MINIFICPP-2629 Clean up minifi-api includes
- Remove duplicate header files with the same content
- Remove proxy header files in core/extension-framework that only
  include minifi-api headers
- Include only the minifi-api headers when implementation is not used
- Fix clang tidy issues

+++ libminifi/src/ThreadedSchedulingAgent.cpp
@@ -30,11 +30,10 @@
 #include <vector>
 
 #include "core/ClassLoader.h"
-#include "core/Connectable.h"
 #include "core/Processor.h"
-#include "core/ProcessContext.h"
-#include "core/ProcessContextBuilder.h"
+#include "minifi-cpp/core/ProcessContext.h"
 #include "core/ProcessSessionFactory.h"
+#include "minifi-cpp/core/ProcessContextBuilder.h"
 #include "utils/ValueParser.h"
 
 using namespace std::literals::chrono_literals;
@@ -87,7 +86,7 @@ void ThreadedSchedulingAgent::schedule(core::Processor* processor) {
   std::vector<std::thread *> threads;
 
   ThreadedSchedulingAgent *agent = this;
-  for (int i = 0; i < processor->getMaxConcurrentTasks(); i++) {
+  for (uint8_t i = 0; i < processor->getMaxConcurrentTasks(); i++) {
     // reference the disable function from serviceNode
     processor->incrementActiveTasks();
